test : cost value, SessionCost class type int -> long

+++ src/main/java/nextstep/courses/domain/SessionCost.java
@@ -3,9 +3,9 @@ package nextstep.courses.domain;
 import nextstep.payments.domain.Payment;
 
 public class SessionCost {
-    private final int cost;
+    private final long cost;
 
-    public SessionCost(int cost) {
+    public SessionCost(long cost) {
         if (cost < 0)
             throw new IllegalArgumentException("비용은 음수 가 될 수 없습니다");
         this.cost = cost;
